+++ src/main/java/org/eclipse/keyple/card/generic/GenericExtensionService.java
@@ -15,11 +15,10 @@ import org.eclipse.keyple.core.common.CommonApiProperties;
 import org.eclipse.keyple.core.common.KeypleCardExtension;
 import org.eclipse.keyple.core.service.resource.spi.CardResourceProfileExtension;
 import org.eclipse.keypop.card.CardApiProperties;
-import org.eclipse.keypop.reader.CardReader;
+import org.eclipse.keypop.genericcard.GenericCardApiFactory;
+import org.eclipse.keypop.genericcard.GenericCardSelectionExtension;
 import org.eclipse.keypop.reader.ReaderApiProperties;
 import org.eclipse.keypop.reader.selection.IsoCardSelector;
-import org.eclipse.keypop.reader.selection.spi.CardSelectionExtension;
-import org.eclipse.keypop.reader.selection.spi.SmartCard;
 
 /**
  * Card extension service providing basic access to APDU exchange functions with a card.
@@ -43,25 +42,13 @@ public final class GenericExtensionService implements KeypleCardExtension {
   }
 
   /**
-   * Creates an instance of {@link CardSelectionExtension}.
+   * Returns an instance of {@link GenericCardApiFactory}.
    *
    * @return A not null reference.
-   * @since 2.0.0
+   * @since 4.0.0
    */
-  public GenericCardSelectionExtension createGenericCardSelectionExtension() {
-    return new GenericCardSelectionExtensionAdapter();
-  }
-
-  /**
-   * Creates an instance of {@link CardTransactionManager}.
-   *
-   * @param reader The reader through which the card communicates.
-   * @param card The initial card data provided by the selection process.
-   * @return A not null reference.
-   * @since 2.0.0
-   */
-  public CardTransactionManager createCardTransaction(CardReader reader, SmartCard card) {
-    return new CardTransactionManagerAdapter(reader, card);
+  public GenericCardApiFactory getGenericCardApiFactory() {
+    return new GenericCardApiFactoryAdapter();
   }
 
   /**